Move serialize functions in separate files.
