+++ runtime/memcache.cpp
@@ -17,9 +17,9 @@ static_assert(TL_ENGINE_MC_GET_QUERY == ENGINE_MC_GET_QUERY, "bad ENGINE_MC_GET_
 #include "runtime/datetime.h"
 #include "runtime/interface.h"
 #include "runtime/math_functions.h"
-#include "runtime/misc.h"
 #include "runtime/openssl.h"
 #include "runtime/regexp.h"
+#include "runtime/serialize-functions.h"
 #include "runtime/string_functions.h"
 #include "runtime/zlib.h"
 #include "server/php-queries.h"
